requirements.txt
fastapi
uvicorn
httpx
python-dotenv
oracledb
